% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_3dpoint.R
\name{d3_3dpoint}
\alias{d3_3dpoint}
\title{3d demo https://gist.github.com/hlvoorhees/5986172}
\usage{
d3_3dpoint(data, ...)
}
\description{
3d demo https://gist.github.com/hlvoorhees/5986172
}
\examples{
{
\dontrun[
d3plot(x = Sepal.Length, y = Sepal.Width, z = Petal.Length, color = Species, data = iris) |> d3_3dpoint()
d3plot(x = cyl, y = gear, z = wt, color = vs, data = mtcars) |> d3_3dpoint()
]
}
}
