% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_line.R
\name{d3_line}
\alias{d3_line}
\title{Create a new D3 line plot}
\usage{
d3_line(data, ...)
}
\arguments{
\item{data}{data in JSON format, created by d3plot() or from another source}

\item{...}{additional, optional arguments}
}
\description{
\code{d3_line()} creates a D3 line plot. It takes data either directly or passed to it from \code{\link{d3plot()}} function.
}
\examples{
\dontrun{
# OK
d3_point(d3plot(x = mpg, y = cyl, color = cyl, id = rownames(mtcars), data = mtcars))
# Better
d3plot(x = Sepal.Length, y = Petal.Length, color = Species, data = iris) \%>\% d3_line()
}
}
\seealso{
\code{\link{d3plot}}
}
