% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_point.R
\name{d3_point}
\alias{d3_point}
\title{Create a new D3 scatter plot}
\usage{
d3_point(data, ...)
}
\arguments{
\item{data}{data in JSON format, created by d3plot() or from another source}

\item{...}{additional, optional arguments}

\item{stroke}{color that is assigned the outside of the point}
}
\description{
\code{d3_point} creates a D3 scatter plot. It takes data either directly or passed to it from \code{\link{d3plot}} function.
}
\examples{
\dontrun{
# OK
d3_point(d3plot(x = mpg, y = cyl, color = cyl, id = rownames(mtcars), data = mtcars))
# Better
d3plot(x = Sepal.Length, y = Petal.Length, color = Species, data = iris) \%>\% d3_point(radius = 5)
}
}
\seealso{
\code{\link{d3plot}}
}
