% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3plot.R
\name{d3plot}
\alias{d3plot}
\title{Create a new D3 plot}
\usage{
d3plot(x, y, data = NULL, ...)
}
\arguments{
\item{x}{the x-axis (independent) variable}

\item{y}{the y-axis (dependent) variable}

\item{data}{the dataset containing the variables}

\item{...}{additional, optional arguments}
}
\description{
\code{d3plot} initializes a d3plot object. It can be used to declare input data frame for a D3 plot and specify various aesthetics as optional input arguments.
}
\details{
\code{d3plot} is used to construct D3 plots. They can be incrementally created
using the forward-pipe operator from the \code{magrittr} package or without this operator. It is strongly recommended to use the forward-pipe operator.
}
\examples{
\dontrun{
d3plot(x = mpg, y = cyl, data = mtcars) \%>\% d3_point()
}
}
\seealso{
\code{\link{d3_line}}, \code{\link{d3_point}}
}

