% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_point.R
\name{d3_point}
\alias{d3_point}
\title{Create a new D3 scatter plot}
\usage{
d3_point(data, css = FALSE)
}
\description{
\code{d3_points()} creates a D3 scatter plot. It takes data either directly or passed to it from \link{\code{d3plot()}} function.
}
\details{
@export
 @param data data in JSON format, created by d3plot() or from another source
 @param css additional css stlying
 @param ... additional, optional arguments

@examples
 \dontrun{
 # OK
 d3_point(d3plot(x = mpg, y = cyl, data = mtcars))
 # Better
 d3plot(x = mpg, y = cyl, data = mtcars) %>% d3_point()
 }
 @seealso \link{\code{d3plot}}
}

